app/requirements.txt
langgraph
langchain
langchain_tavily
motor
fastapi
uvicorn
psycopg2-binary
mysql-connector-python
langgraph.checkpoint.mongodb
dotenv